requirements.txt
bleak==0.14.3
paho-mqtt